+++ wallet-connect-test-bench/front-end-with-dapp-library/src/Root.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 
+import { WithWalletConnector } from '@concordium/react-components';
 import Main from './Main';
+import { TESTNET } from './constants';
 
 /**
  * Connect to wallet, setup application state context, and render children when the wallet API is ready for use.
@@ -9,7 +11,7 @@ export default function Root() {
     return (
         <div>
             <main>
-                <Main />
+                <WithWalletConnector network={TESTNET}>{(props) => <Main {...props} />}</WithWalletConnector>
             </main>
         </div>
     );