+++ wallet-connect-test-bench/front-end-with-dapp-library/src/reading_from_blockchain.ts
@@ -1,6 +1,5 @@
 import { toBuffer, JsonRpcClient, deserializeTypeValue, deserializeReceiveReturnValue } from '@concordium/web-sdk';
 
-import { WalletApi } from '@concordium/browser-wallet-api-helpers';
 import {
     CONTRACT_NAME,
     CONTRACT_INDEX,
@@ -151,8 +150,8 @@ export async function getValue(rpcClient: JsonRpcClient, useModuleSchema: boolea
     }
 }
 
-export async function view(client: WalletApi) {
-    const res = await client.getGrpcClient().invokeContract({
+export async function view(rpcClient: JsonRpcClient) {
+    const res = await rpcClient.invokeContract({
         method: `${CONTRACT_NAME}.view`,
         contract: { index: CONTRACT_INDEX, subindex: CONTRACT_SUB_INDEX },
     });
@@ -173,29 +172,6 @@ export async function view(client: WalletApi) {
         return JSON.stringify(state);
     }
 }
-
-// export async function view(rpcClient: JsonRpcClient) {
-//     const res = await rpcClient.invokeContract({
-//         method: `${CONTRACT_NAME}.view`,
-//         contract: { index: CONTRACT_INDEX, subindex: CONTRACT_SUB_INDEX },
-//     });
-
-//     if (!res || res.tag === 'failure' || !res.returnValue) {
-//         throw new Error(
-//             `RPC call 'invokeContract' on method '${CONTRACT_NAME}.view' of contract '${CONTRACT_INDEX}' failed`
-//         );
-//     }
-
-//     const state = deserializeTypeValue(toBuffer(res.returnValue, 'hex'), toBuffer(VIEW_RETURN_VALUE_SCHEMA, 'base64'));
-
-//     if (state === undefined) {
-//         throw new Error(
-//             `Deserializing the returnValue from the '${CONTRACT_NAME}.view' method of contract '${CONTRACT_INDEX}' failed`
-//         );
-//     } else {
-//         return JSON.stringify(state);
-//     }
-// }
 
 export async function accountInfo(rpcClient: JsonRpcClient, account: string) {
     return rpcClient.getAccountInfo(account);