+++ wallet-connect-test-bench/front-end/CHANGELOG.md
@@ -1,7 +1,5 @@
 ## Unreleased changes
 
-- Add `deploy`/`initialize` testing scenarios
-
 ## 1.0.0
 
-- Initial wallet test bench with dApp libraries
+- Initial wallet test bench without dApp libraries